requirements.txt
asgiref==3.11.1
beautifulsoup4==4.12.3
Django==6.0.4
lxml==6.1.0
python-docx==1.1.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
gunicorn
whitenoise
dj-database-url
psycopg2-binary
